+++ player.c
@@ -17,7 +17,7 @@
 /**
  * Lê coordenadas e direção para posicionar navio
  */
-void scan_ship_position(uint16_t *row, uint16_t *col, bool *direction)
+void scan_ship_position(uint8_t *row, uint8_t *col, bool *direction)
 {
 	uint16_t tmp_row, tmp_direction;
 	char tmp_col;
@@ -32,7 +32,7 @@ void scan_ship_position(uint16_t *row, uint16_t *col, bool *direction)
 /**
  * Lê coordenadas de tiro
  */
-void scan_shot_position(uint16_t *row, uint16_t *col)
+void scan_shot_position(uint8_t *row, uint8_t *col)
 {
 	uint16_t tmp_row;
 	char tmp_col;
