+++ player.h
@@ -1,7 +1,7 @@
 #ifndef __PLAYER_H__
 #define __PLAYER_H__
 
-void scan_ship_position(uint16_t*, uint16_t*, bool*);
-void scan_shot_position(uint16_t*, uint16_t*);
+void scan_ship_position(uint8_t*, uint8_t*, bool*);
+void scan_shot_position(uint8_t*, uint8_t*);
 
 #endif
